# improve performance of find_stack_level